Do not expose converted structs to C code, remove C string conversions
