+++ src/c/mod.rs
@@ -1,0 +1,67 @@
+use libc;
+
+use std::ffi::CString;
+
+use log;
+
+pub use self::config::*;
+pub use self::disk::*;
+pub use self::filesystem::*;
+pub use self::installer::*;
+pub use self::partition::*;
+pub use self::sector::*;
+
+mod config;
+mod disk;
+mod filesystem;
+mod installer;
+mod partition;
+mod sector;
+
+/// Log level
+#[repr(C)]
+#[derive(Copy, Clone, Debug)]
+pub enum DISTINST_LOG_LEVEL {
+    TRACE,
+    DEBUG,
+    INFO,
+    WARN,
+    ERROR,
+}
+
+/// Installer log callback
+pub type DistinstLogCallback = extern "C" fn(
+    level: DISTINST_LOG_LEVEL,
+    message: *const libc::c_char,
+    user_data: *mut libc::c_void,
+);
+
+/// Initialize logging
+#[no_mangle]
+pub unsafe extern "C" fn distinst_log(
+    callback: DistinstLogCallback,
+    user_data: *mut libc::c_void,
+) -> libc::c_int {
+    use DISTINST_LOG_LEVEL::*;
+    use log::LogLevel;
+
+    let user_data_sync = user_data as usize;
+    match log(move |level, message| {
+        let c_level = match level {
+            LogLevel::Trace => TRACE,
+            LogLevel::Debug => DEBUG,
+            LogLevel::Info => INFO,
+            LogLevel::Warn => WARN,
+            LogLevel::Error => ERROR,
+        };
+        let c_message = CString::new(message).unwrap();
+        callback(
+            c_level,
+            c_message.as_ptr(),
+            user_data_sync as *mut libc::c_void,
+        );
+    }) {
+        Ok(()) => 0,
+        Err(_err) => libc::EINVAL,
+    }
+}
